related objects while delete should have filter on tenant_id, use django ORM filters instead of raw sql
